requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
lxml==4.4.1
motor==2.0.0
multidict==4.5.2
pymongo==3.9.0
python-dateutil==2.8.0
python-dotenv==0.10.3
six==1.12.0
soupsieve==1.9.3
typing-extensions==3.7.4
yarl==1.3.0
